requirements.txt
redis
click
semantic_version
PyYAML